[FEAT] Implementation of slivers to increase performance on big lists
